Changed types of Camera parameters (double -> float). 1. Changed types of Camera parameters (double -> float).

+++ Engine/GraphicEngine/Managers/ViewManager/ViewManager.h
@@ -59,7 +59,7 @@ namespace MatWiecz
             
             ViewManagerRetVal UnregisterCamera(unsigned int cameraId);
             
-            ViewManagerRetVal UpdateProjection(double aspect);
+            ViewManagerRetVal UpdateProjection(float aspect);
             
             ViewManagerRetVal PerformViewTransformation();
